+++ internal/manifest/defaults_tags.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"strconv"
 
-	"github.com/shazow/virtle/internal/units"
+	"github.com/shazow/virtle/units"
 )
 
 var durationType = reflect.TypeOf(units.Duration(0))